requirements.txt
Flask==0.12.2
line-bot-sdk==1.8.0
beautifulsoup4==4.7.
urllib3==1.24.1
soupsieve==1.6.1
pytz==2019.3
